+++ requirements.txt
@@ -5,7 +5,6 @@ django-compressor==2.3
 django-compressor-toolkit==0.6.0
 django-filter==2.1.0
 django-guardian==1.5.1
-django-heroku==0.3.1
 django-webpack-loader==0.6.0
 djangorestframework==3.9.4
 gunicorn==19.9.0
