+++ visualizer/static/visualizer/components/App/App.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import AppTest from "./App.jsx";
-import 'bootstrap/dist/css/bootstrap.css'
+import Index from "../Index";
 
 /**
- * Register a Delivery to a selected company
+ *
  * @param props
  * @return {*}
  * @constructor
@@ -13,12 +12,11 @@ class App extends React.Component {
     render() {
         return (
             <div>
-                <AppTest/>
-                {/*<TestStuff/>*/}
+                <Index/>
             </div>
 
         )
     }
 }
 
-ReactDOM.render(<App/>, document.getElementById("root"));
+ReactDOM.render(<App/>, document.getElementById("app"));
