--- visualizer/static/visualizer/components/Root/App.jsx
@@ -1,24 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Routes from "../Routes/Routes.jsx";
-// import succ from '../../media/successful.mp3'
-
-/**
- * Register a Delivery to a selected company
- * @param props
- * @return {*}
- * @constructor
- */
-export default class AppTest extends React.Component {
-
-
-
-    render() {
-        return (
-            <div>
-                <Routes/>
-            </div>
-        )
-    }
-}
-
