+++ visualizer/templates/root.html
@@ -17,7 +17,7 @@
   {#    DECLARE THE STATIC URL TO BE USED WITHIN THE APPLICATION WHEN LOADING SERVER SAVED PERMANENT MEDIA (Not much for this project so it's enough#}
     let STATIC_URL = "{% get_static_prefix %}";
 </script>
-<div id="root">
+<div id="app">
 
 </div>
 <script src="{% static 'main.js' %}"></script>
